+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2026-04-25
+
 ### Fixed
 
 - iPhone notch area now shows dark header color in Safari (body background set to header color, main/footer have their own light background)